fix: home page server component
